[TASK] Changed inline content relation to 1:n
To work around the issue that TYPO3 will not copy the inline content
elements when a news record is copied because they are attached to the
news in an mm relation the relation type was changed.

The inline content elements and the teaser content elements are now
related in a 1:n realtion to the news record. Therefore two news
fields have been added to the tt_content table.

Additionally and Extbase command controller was added for updating old
tt_content records.